添加判断脚本是否已经存在

+++ moa/plugins/testlab/stf_runner.py
@@ -118,7 +118,8 @@ def run(addr, moa_script, utils_dir=""):
     import shutil
     import subprocess
     filename = os.path.basename(moa_script)
-    shutil.copytree(moa_script, filename)
+    if not os.path.exists(filename):
+        shutil.copytree(moa_script, filename)
     utilfile = os.path.join(utils_dir, "utils.py")
     p = subprocess.Popen(["python", "-m", "moa.airtest_runner", filename,
         "--setsn", addr, "--log", "--screen", "--utilfile", utilfile
